backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
motor==3.6.0
pymongo==4.9.2
pydantic==2.6.4
pydantic[email]==2.6.4
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
PyJWT==2.8.0
python-dotenv==1.0.1
python-multipart==0.0.9
httpx==0.27.2
opencv-python-headless==4.10.0.84
ultralytics>=8.2.0,<9.0.0
imageio-ffmpeg==0.6.0
